TIP/requirements.txt
opencv
numpy
matplotlib
pillow
skimage